Small script to pack strings into fixed bytes
packer.pack(data, length) will flatten data, which can be an arbitrarily-nested
block of lists (i.e., just a list of words, or a list of list of words, or
whatever, you can even mix it up), and then pick words out of that flattened set
in order until it finds one that goes over the length limit.

If you run it with the optional argument pack=True, then it will just keep
adding words, ignoring ones that don't fit, until it runs out of words.